refactor: move signin component code into separate utility file
